Add shouldWait and isComplete to useCachedDomainContactDetails
`shouldWait` is an argument that can be used to delay fetching contact
details (and therefore also to delay filling the contact form).

`isComplete` is a return value to let the caller know that the contact
form has been auto-completed (or at least completion has been
attempted).

Because this changes the arguments, this is a breaking change.