Move contact step into custom step component
We want to call `useCachedDomainContactDetails` above `WPContactForm` so
that we can determine when the contact form auto-fill validation is
in-progress and silence the validation errors during that time.

However, to have access to step completion the hook must be called
inside `CheckoutStepGroup`, and we will need its return value for the
`isCompleteCallback` prop of the contact form step.

Therefore, we must put a wrapper component inside `WPCheckout` to call
the hook and use its return value for the contact form step.