Checkout: always use selected site for checkout masterbar UI (#69935)
* Use currentSelectedSiteSlug for checkout masterbar slug

Previously this used the primary site as a fallback when the selected
site was undefined but that makes the masterbar for siteless checkout
(where an undefined site is expected) behave unpredictably.

* Use selected site for isJetpackNotAtomic in checkout masterbar

This prop is only used for the checkout masterbar, which needs to
support undefined as a selected site. Previously this calculated
`isJetpackNotAtomic` for the selected site or the primary site if none
was selected, but that causes unpredictable behavior for siteless
checkout.