Fix showing licensed quantity during checkout for Jetpack Search
Before we introduced new pricing for Jetpack Search,
there used to be a subtext under checkout line item
with tier name, for example "Up to 100 records".
This is not working with new pricing.

In this commit, I'm changing it so that we show text
like "Up to 20k records and/or requests per month"