Add support for a monthly-to-yearly domain nudge
A monthly plan, getting a JITM nudge to upgrade to annual to get a free
domain, would need to have the plan step show yearly plans. Monthly
plans don't have a free domain credit.

This change will ask monthly plans to switch to an annual plan to get a
free domain, if the user clicks on a link like this:

/domains/add/YOUR-TEST-SITE.wordpress.com?domainAndPlanPackage=true

Sites on a 2-year plan or on Free should be unaffected by the change