Fix some sign-up flows which put user step last
The domain, add-domain and with-theme sign-up flows put the user step
(login/signup) last. If a user already had an account they were prompted
to login, once logged in, they were sent back into the sign-up flow which
gots comprehensively stuck.

This change has fixed the problem by resurrecting some code implemented
in #44172 which removed and readded the in-progress step so it would
reprocess with the signup flow.

Fixes #70091