Display appropriate plan remove modal for marketplace themes
When a user has both an atomic plan (business or higher) in their cart
along with one or more marketplace theme subscriptions, we want to
display a modal warning them the themes will be removed when they remove
the business plan from the cart.

The 'plan with marketplace dependencies' modal copy already says what we
need so this is a small change to piggy back on that modal.