Add key icon to material-design-icons package
Used instructions in README to add the icon to the package.

The icon ID, used to look up the icon inside the sprite map, includes
the icon size. Since we eventually want the icon to be 32px that means
we unfortunately need to add "32px" in the SVG file name; even though
it's still a 24px SVG under the hood.
This is how the cloud icon we currently use works.

Original icon comes from:
https://fonts.google.com/icons?selected=Material+Icons&icon.query=key&icon.style=Outlined&icon.category=communication