feat: Disable Jetpack app banner Android deep links
Remove the `url` parameter as a workaround for the WordPress to Jetpack
app migration experience not launching when navigating to a deep link.
Removing the `url` parameter means deep links from these banners will
not navigate the user to the correct screen after the app launches. We
intend to reinstate this `url` parameter once the Jetpack app is updated
to launch the migratoin experience alongside deep links.