<SitesDisplayModeSwitcher> supports keyboard navigation
Because the display mode toggle isn't a real radio group it means it's
up to us to implement all the expected behaviour a radio group has.

This'd be easier to implement if we used real <input type="radio">
elements, but since we want the buttons to be styled like the <Button>
component we need the manual implementation.

This W3C resource defines how radio groups are supposed to behave.
https://www.w3.org/WAI/ARIA/apg/patterns/radiobutton/

In this case I've gone with the method where only one of the buttons is
focusable at one time by setting the tabindex.