Change behaviour of dotorg theme nudge
Link the dotorg theme nudge displayed to sites that are not atomic
capable straight to the checkout for the business plan.

Include a redirect so that post-purchase, the user is returned back to
the themes page so they can then activate the theme.

Make the nudge wording clearer so that what you are buying is more
explicit.