Fix CoBlocks fonts issue for simple sites (#72679)
CoBlocks typography support is adding an extra style rule to blocks when
using inserting patterns or pasting content with custom fonts. This happens
even on sites that are not using any CoBlocks functionality. It surfaces to users
as an "Invalid content" error when inserting content, and when previewing
patterns.

This change removes the filter that does that - we disable controls for
typography support anyway.