Limited Global Styles: Add E2E test (#71333)
Adds a new E2E test to check that, when Global Styles are limited, the upgrade modal and the upgrade pre-save notice show up in the Site Editor.

Global Styles became a paid feature (available on Premium plans and above) for sites created after Dec 15, 2022. See p7DVsv-fY6-p2.

That means that Free and Personal sites created after that date would have Limited Global Styles. We basically allow them to try Global Styles, but any custom style is not loaded in the frontend.

As part of the Limited Global Styles flow, we show upgrade banners in the Site Editor. However, those are fragile because they rely on a CSS selector to manipulate the DOM and inject those banners, but Gutenberg can change the DOM at any point without further notice.

So with this test, we hope to catch those regressions when new Gutenberg versions are released to WPCOM so we can update the Limited Global Styles implementation with new CSS selectors.