Add filename suffix option to Webpack RTL plugin
Currently, the Webpack RTL plugin always appends .rtl.css to the filename. This is not always desirable, as it limits the plugin usage in other projects. In the original plugin created by Romain Berger there was an option to customize the filename, so this commit adds an option to the plugin to allow the user to specify a custom suffix.