WooExpress trial checks that WooCommerce is installed before completing
Previously, we redirected to `/home` as soon as the Atomic transfer was
complete. This lead to issues because, sometimes, WooCommerce wasn't
finished installed before the redirect so we'd end up on the main
dashboard instead of the WooCommerce launch checklist.

This adds a new step where we wait for WooCommerce to finish installing
before we redirect.