Use the ONBOARD_STORE for the list of plugins to verify
In the `siteCreationStep` at the beginning of the flow, we pass the list of plugins to wait for installation of to the ONBOARD_STORE.

The `waitForPluginInstall` step then uses that list rather than having
special cases in the step for different flows.

If there are no plugins to check, the step is effectively skipped.