Change tooltip in onboarding for virtual themes (#73423)
Premium themes and virtual themes made of style variations of premium
and non-premium themes require the Premium or above plan to use. These
themes and v-themes display a 'Premium' badge that has a tooltip
explaining what premium means.

For virtual themes made of style variations on sites that are on a lower
plan than premium this tooltip will now be more relevant.