Add context when adding global-stylesstickers (#74017)
Add a link to the P2 describing the global styles gating stickers when
adding them.

The stickers can be added when superadmins visit the sites blog report
card, this shows up in the audit log and may be alarming. Adding this
link means that people who notice this can read more about it and gain
some reassurance.