Update Connect Domain screens to allow for Domain Connect (#73507)
Refactor the Connect a domain code to actually check if Domain Connect is supported and our hosting template is supported and use that by default to map a domain to WordPress.com. The backend code needed for this to work is temporary disabled so nothing should change for the customer as of now.