Add support links for core/group and core/navigation
This PR adds a learn more link for the core/group block (and its variations - row and stack) and for the core/navigation block.

For core/group, we have a special handling because the block description is generated based on the variations property, since the row and stack blocks are variations of the group.

To achieve this, a new map was introduced that takes into account the variation property.