Update: Remove checking for production_blog_id in SitePlan compoenent
Originally, this check was added for safety reasons. If for example
the site had incorrectly only the staging sticker and not the production_blog_id
this would show incorrect results.
Rely only to staging sticker for now until we decide what we want to do
(show production site title or not). This change simplifies the component
for now.