Do not cache log results by scroll_id, use a dummy page index param instead
The log requests are not idempotent. The scroll ID actually represets a
query state that lives on the server for 10 minutes. We can't use the
scroll ID as a mechanism for fetching the next page since it will be a
consistent value across all pages.

The page index is now used to store the log entry in ReactQuery. The
page index isn't actually part of the request body, but it acts as a
good key for persisting the logs.