Fix timezones so they are always displayed in "site" time
The idea is to pass timestamps around as Moment objects and not worry
too much about what timezone they're in until right before we need to
display them, and then we convert at the last minute to ensure things
are correct.

The times saved to the URL as query params are in UTC time for maximum
URL portability. There is no need to convert them to UTC before calling
`.unix()` because "unix time" implies UTC.