Fix untranslated text on /stats/ads/ page.
This is a combination of two problems:
- The 'Ads' navigational item always appeared in English
- Ad chart tabs containing text like 'Ads Served' and 'Average CPM' can appear untranslated.

The first issue was fixed by wrapping the text, and getters with translate() calls.

The second issue is more beneign. Although the strings are wrapped in translate() calls,
they'll only appear translated depending on the locale settings at the time of loading.
Using Javascript getters allow the text to appear translated at the time the label is queried
by the UI.