i18n: Percent discount not localized in upsell during cancel flow
In the cancel flow, the plan upsell screen contains a discount,
but the percent sign and the discount amount are hard coded and not localizable.
In some languages, there needs to be a space between the amount and the percent sign.
And in some languages, the percent sign should be in front of the amount.