Site Editor: un-iframe, again. Revert the revert.
Reverts #76106

The CORS-issue for wpcom mapped domains was resolved in D108844-code
Because we serve the Site Editor from the `*.wordpress.com` subdomain, we needed to add that subdomain to an Access-Control-Allow-Origin header for the frontend when it is mapped to a custom domain.
Because the iframe was directly passing the appropriate `postId` and `postTemplate` query args to the Site Editor, it hid that longstanding issue long enough for us to forget about it.