Staging sites: fix staging creation navigation bug
When the user creates a new staging site, and move back and forth to
the page while the staging creation is in progress, faces strage issues
like:

- The button to create a site becomes visible again
- When the simple staging site has been created, an error message is
  displayed instead of the progress bar.

This pr fixes the above issues.

NOTE: This doesn't fix the case where the user reloads the page instead.
In that case we have implemented a lock mechanism in D109053 that
prevents the user to create a new site.