Support chat using Zendesk Messaging (#76777)
We're switching from Happy Chat to Zendesk Messaging for live chat support. This PR is big, but it's also the smallest I could get given the scope of the project. Any cleanup of Happy Chat code, old contact forms, etc. will be done as a follow-up. I want to keep this PR as easy to revert as possible if needed.


    Disable Happy Chat feature flag and any availability checks that trigger connections to the websocket.
    Load the Zendesk Messaging script instead.
    Authenticate using already established endpoint (just with type=zendesk).
    Check availability using the /help/messaging/is-available endpoint.
    Pass the initial chat data using Zendesk user fields.
    Use the support history to figure out if the user has ongoing conversations or not (thankfully, a Messaging conversation is a "ticket"!).
    Additionally, make sure the Zendesk launcher is visible if any unread messages are reported by Zendesk's APIs directly. This is a good safeguard and makes it more dynamic.
    If the browser is blocking requests to Zendesk, show a notice and link to relevant support doc.