Jetpack Connect: Remove dependency on Happy Chat
Part of sunsetting Happy Chat: pdm0RZ-BU-p2

I'm not sure what was the original intent here - JPOP does not do chat,
so at best this would have connected to WPCOM HEs. Plus, it seems some
of these flows are when user is logged out - which did not work with
Happy Chat (which requires WPCOM user). All in all, I think it's best to
go back to the fallback contact URL and then we can see if any changes
are needed (like showing Help Center).