Add `getThemeType` selector (#77550)
In order to reuse the same component for rendering the theme type badges in the design picker and in the theme showcase (https://github.com/Automattic/dotcom-forge/issues/2438) we need a single way to infer the theme type.

This PR introduces a new `getThemeType` selector to centralize that logic rather than duplicating it across the codebase. The new selector will be used by a follow-up PR which will refactor how the theme type badges are rendered.

For now, the `/with-theme` flow has been refactored to use the new selector.