Add `canUseTheme` selector (#77551)
In order to reuse the same component for rendering the theme type badges in the design picker and in the theme showcase (https://github.com/Automattic/dotcom-forge/issues/2438) we need a single way to infer whether the site has needed plan to activate a certain theme type.

This commit introduces a new `canUseTheme` selector to centralize that logic rather than duplicating it across the codebase. The new selector will be used a by a follow-up commit which will refactor how the theme type badges are rendered.