Boost the specifity of the badge premium badge
The css for the design picker's PremiumBadge is being and is
out-specifying the CSS here. This results in an incorrectly
formatted tooltip for marketplace and premium themes but not for Woo.

I've increased the specifity here as a work-around but we need to
carefully test to ensure that the problems addressed in #77813 do not
reoccur.