Plans 2023: Toggle layout based on screen size
Fix mobile view style

Mock `useIsLargeScreen` hook in the `PlansFeaturesMain` test file.

refactors the plan-features-2023-grid module to use a new hook (`use-get-grid-size`) instead of the `use-is-large-screen` hook. The new hook provides the grid size (small, medium, or large) instead of a boolean value, and takes into account whether the sidebar is collapsed or not.

[plan-features-2023-grid] Add useResizeObserver to handle grid resizing