Improve pre-cancellation chat message with additional info
The product name is in the user's locale, so it's not always clear what
it's referring to. Looks like the product slug is the closest we can get
to an always English name.
Additionally, let's make sure to reference the domain name if we're dealing with a domain subscription (registration or connection).

Fixes p8wKgj-4gB-p2#comment-20528