Refactor download button and Tracks
Given that we have the URL for downloading WordPress when rendering the
file info card, we are now using the href attribute instead of the
onclick to handle the download.

This commit also includes a refactor of the Tracks tracking action to
also be triggered when downloading a database table and WordPress.