Add support for a `/<locale>/plans` route
In email and support documents, we often link to
wordpress.com/plans. Logged in users are directed
to the plans in Calypso, but logged out users
are redicted to an English pricing page (wordpress.com/pricing)

This commit adds support for a plans page that includes the locale.
We'll update the localizeUrl function to add the locale to this page
and redirect the user accordingly