Hardcode GS availabilty to personal plan
Global styles is now part of the personal plan rather than being in an
A/B test between Personal and Premium. To clean-up the code after the
test the PHP APIs will defer to the Features API as to whether the user
has a sufficient plan, and will hardcode that the minimum plan required
for GS on this site is the personal plan.

Future clean-up diffs will remove the need to indicate the minimum plan.