Remove GS experiment & support treatment sites
Global Styles is back to being part of the Premium plan rather than
being in an A/B test between Personal and Premium. To support sites who
had bought the personal plan during the test due to being on that side
of the experiment a sticker is checked to see if they have eligibility
for global styles.