Ensure Help Center does not disappear on view change (#79742)
I've run into this while working on #79741.

Looks similar root issue to
https://github.com/Automattic/wp-calypso/pull/78236.
But with different side-effect - since `handleClose` changes on every
view change, Help Center gets removed from DOM (and `handleClose` gets
called to close it). Then it gets immediately re-added to DOM. But it
does not get re-opened, of course. The result is that Help Center
disappears if you switch views in Calypso, which is definitely not how
it's intended to work.

This fixes the root issue, making sure that `handleClose` does not
change on view change, thanks to useCallback.