Show support links in Help Center instead of modal
As suggested in pbvpgB-3o7-p2

If a user clicks on a support doc link in Calypso, a new modal (or tab) is opened with it. This takes the user away from the context they’re need that information the most. Thanks to Vertex's work on Help Center, we can instead show those pages the Help Center. This will keep the user in the Calypso context, while giving them a better UX. The floating window of Help Center can be moved around and stays as they navigate Calypso and, for example, follow the steps on the support doc.

I've used the existing InlineSupportLink component that is used across Calypso (though, not consistently). It has been introduced in 2018 in #26347, and fits perfectly our needs.