Extract the dialog code into <DialogPaidPlanIsRequired> in preparation of introducing a new .blog + Free domain dialog.
Introduce `<DialogPaidDomainAndFreePlan>`

Dedupe the props of the two dialogs.

Extract the shared logic of getting a suggested plan into
<GetSuggestedPlanSection> component

Edit the description of `DialogPaidDomainAndFreePlan`

Add the minimal version of the hool `useIsCustomDomainAllowedOnFreePlan`

Differ the allowed actions and copies based on
`isCustomDomainAllowedOnFreePlan`

Emphasize that the custom domain will be a redirect in the feature list

Customize the copy and CTA of the .blog + Free plan dialog.

GetSuggestedPlanSection -> SuggestedPlanSection

Remove the redundant trailing space.

Remove the comment that is soon to be meaningless.

Move `useIsCustomDomainAllowedOnFreePlan` out to <PlanFeaturesMain/>

Separate `replacePaidDomainWithFreedomain` into `removePaidDomain` and
`setSiteUrlAsFreeDomain`, so the paid domain can be conditionally
removed.