Domain Management i1: Creates initial `<DomainsTable>` component (#80539)
* Add boilerplate for new @automattic/domains-table package

The components and hooks needed for the domains table will be in this
package. Add the components to a seperate package will aid us in our
goal to have the table be embeddable in different locations. This is
important because the IA for domains is in flux.

* Add <DomainsTable> to storybook

Creates the basic boilerplate of a domains table with initial styling.

* New hooks for fetching row data for <DomainManagement>

* Render the new domains table at /domains/manage when flag is enabled