Add Individual Subscriber Stats (#80614)
* Add individual subscriber stats cards, currently behind a feature flag.

* Add individual subscriber stats cards, currently behind a feature flag.

* Refactor subscriber stats open and click rates calculation

This commit removes repetitive rate calculation code in the subscriber stats component (open and click rates) and replaces it with a reusable hook `useSubscriberStatsRate`. It increases maintainability, since changes to the rate calculation now only need to be implemented in one place. One icon file `chart.tsx` was removed as it's replaced by using icon from WordPress icons directly. A dedicated type `SubscriberStats` was moved into a central types file to be re-used across components.

* Enable individual subscriber stats in production

* Add function to format subscriber rates

Added a function to handle formatting of subscriber stats rates in `useSubscriberStatsRate.ts`. The function ensures that integer rates are returned as integers and decimal rates are rounded to 1 decimal place. This change provides a consistent data type for the returned rates and improves readability.