Switch to getting the experiment variation from the backend
useExperiment cannot be used as-is in Help Center, and fetching the
variation from the front end here requires significant development.

Here, we work around that by getting the variation from the backend while
getting the user chat availability.

Since we don't show the screens affected by the experiment until the chat
status is loaded, we also don't need to make any further changes to
handle the loading state