Display an UpgradeModal on dotorg themes
On a community (dotorg) theme sheet display a button to activate the
theme rather than just a banner prompting purchase of the business plan.

This button will display a modal (if necessary) extolling the virtues of
a business plan and explaining that it is a prerequisite to using the
theme. This is the same behaviour as partner, woo themes etc but with
slightly different text.

To facilitate this the ThemeUpgradeModal uses the `getCanonicalTheme`
selector, which means the `useThemeDetails` hook is obsolete and no
longer used.