Add feature flag for newsletter categories
A new hook "useNewsletterCategoriesFeatureEnabled" has been added to check if the newsletter categories feature is enabled for a particular site. The hook uses data fetched from the "blog-stickers" endpoint. This was implemented with the intent of providing more controlled feature access. Updates have been made in corresponding files to utilize this new hook. Relied on react-query for data fetching. Added unit tests for the new hook as well.