2023 Pricing Grid: Use smallest unit for currencies in the plans pricing grid (#81537)
* Use smallest units for all pricing in 2023 plans pricing grid

* Fix currency overflow issues

Previously, useIsLargeCurrency was only evaluating  prices individually. It, however, wasn't considering when discounted and undiscounted prices were displayed on the same line. Because of this, we were seeing overflow in the plans comparison grid for AUD, CAD, CHF, etc.

To remedy this, we set a threshold and now evaluate the combined discounted and undiscounted prices for a given plan to decide whether or not a currency is large for its given display

To be clear, this is not an ideal solution. We should explore CSS containers for a more exact way to handle price formatting