"Update address and improve map provider handling in e2e tests
The address used in the end-to-end test for Jetpack Map block was edited to fix a potential location inaccuracy issue.
The code for handling different map providers was also updated to accommodate mapkit's unique rendering mechanism. A strategy will be devised in the future for selecting elements within mapkit's shadow DOM."