Show unmodifiable field warning when only one domain is selected
We weren't showing the unmodifiable field warning message if there was
only one domain selected. That'd make sense if the unmodifiable fields
were also disabled, but they're not because of the bulkEdit prop.

I've fixed the issue by always showing the unmodifiable field warning
regardless of how many domains there are. Another option would have been
to only set bulkEdit=true when there was more than one domain selected.

I decided to go with this option because it's one fewer special modes
for the page to be in. But I can see how it might be nicer to go the
other way too.