Add user-content-links to email stats
In the previous implementation, we only tracked internal links for email statistics. However, there is also a need to include and track user-content-links.

The changes introduced in this commit ensure that user-content-links are also fetched along with ‘link’ stats. Thereafter, these user-content-links are parsed and included in the overall email stats. The code now includes checks for the existence of both internalLinks and userContentLinks. It also includes the sorting of the 'mappedLinks' array..