Email stats: Add user-content-links to email stats (#81882)
* Add user-content-links to email stats

In the previous implementation, we only tracked internal links for email statistics. However, there is also a need to include and track user-content-links.

The changes introduced in this commit ensure that user-content-links are also fetched along with ‘link’ stats. Thereafter, these user-content-links are parsed and included in the overall email stats. The code now includes checks for the existence of both internalLinks and userContentLinks. It also includes the sorting of the 'mappedLinks' array..

* Refactor parseEmailLinksData for array validation

The function parseEmailLinksData has been updated to provide default values for the input parameters. This change removes the need for an early function exit when the input parameters are not arrays. The validation check for the input parameters has been moved for each variable separately to correctly handle the scenario when one of the parameters is not an array. This refactor greatly simplifies the function logic and improves its robustness.