Lazy Images: add notice to settings UI preparing for deprecation
The Lazy Images feature will be deprecated in Jetpack 12.8, scheduled for release in November. When that happens, we want the UI to disappear.

Until then, we want to:

- warn folks that use the feature today that they can disable it.
- stop folks from enabling it if it is not enabled on their site today.

This logic does not take Gutenberg into account, but that is something that will be handled in the Jetpack plugin itself, in https://github.com/Automattic/jetpack/pull/33208/
-> the module will get disabled, thus updating the UI in Calypso in turn.