Storage Add Ons:  Implement dropdown checkout functionality (#81340)
* Add dropdown functionality to storage add-on upsell

* Update plans-ecommerce-fulfilled step

* Always show default storage option slug for plan

As it exists today, storage labels show the default amount of storage in the pricing grid, even if a storage upgrade add-on is purchased. We ensure that we continue to do so if the storage add-on dropdown is not displayed

* Support multiple checkout cart items in buildUpgradeFunction

* Fix various bugs and units tests

Revert isEmpty change

Remove unnecessary check for cartItem

Fix pro plan upgrade button

Fix plan step unit test

Fix typescript errors

* Don't fetch add-on transaction history in onboarding

* Build cartItems in the plan features grid

* Fix case for selectedStorageOptionForPlans type

* Move add-on transaction filter into block scope

* Disable billing transaction query in onboarding

* Fix upgrade click storage add-on data retrieval

* Add named tuple destructuring for clearer tests

* Update inline comment about check for plan free slug

* DRY getPlanCartItem

* Handle cartItems in stepper flows

* Fix stepper onSelectPlan type error

* Fix domain flow redirect error

There was a cartItem dependency for the site-picker step that needed to be updated to cartItems. On closer inspection, however, it appears as if the site-picker step doesn't need the dependency at all. The logic within the site-picker component makes no indication of needing it, so I remove the cartItem dependency instead.

Additionally, the plans step in the domain flow was being skipped under certain circumstances, so we update the provided cartItem dependency to cartItems in those circumstances as well.

* Remove unnecessary selectedStorageOption type

* Add inline TODO note about deprecation of cartItem