Ignore empty behaviours when nudging for global styles
This change fixes a problem where the front-end of sites displayed an
upgrade nudge for global styles in the front-end of a site but did not
display a nudge in the editor. This happened because some sites had an
empty 'behaviours' key in the custom styles. Behaviours was in an
earlier Gutenberg release but has been removed (and a migration path
provided) while it is worked upon further.

The fix works by ignoring the 'behaviours' key if the value is empty.