Allow multiple ThemeCollection on one page (#82467)
Previously only the first ThemeCollection on each page worked as a
carousel. With this change all of the ThemeCollections work as
carousels.

One behaviour which may be a little...surprising is that keyboard
navigation will scroll all ThemeCollection that are in the current
viewport, not just the one which has focus.