Fix with theme signup flow for partner themes
Instead of using the page function to redirect the user to checkout, we are now using the callback approach that's used in other addToCart flows.

This is needed in order for the checkout page to be properly displayed. Previously, the user was pointed to log-in again even though they just signed up for an account on WP.com