Added userId to useSubscribedNewsletterCategories function (#83522)
This commit includes changes to utilize userId in subscribedNewsletterCategoriesData function. The userId is now being used in useSubscribedNewsletterCategories function and related test cases have been added as well. This userId inclusion is necessary and helpful to handle the user data more effectively. The previous version could lead to wrong responses in cases where the subscriptionId is not available but the data is still needed to be fetched by user.