Switch the default category for build intent
The 'Promote myself or business' and 'Other' goals in onboarding are
collectively the 'build' intent. On the design picker during onboarding
this was defaulted to the 'Portfolio' design category since #80751.
Staff feedback since then has been consistently against this.

This change will use the 'Business' category rather than the 'Portfolio'
category