Hide domain checkbox placeholder when there are no selectable domains
The DomainsTable placeholder sometimes looks messy during loading
because we show a placeholder for the checkbox when the internal table
state has `canSelectAnyDomains === false`. This internal state (even
though it's intermediate loading state) controls the CSS grid layout,
and therefore if the placeholder renders a cell for the checkbox it
won't align with the grid layout.

Quick fix is to have the placeholder match the internal table state of
`canSelectAnyDomains` even when it's just intermediate loading data.

A better solution for loading might be to show a consistent loading
experience that doesn't depend on intermediary state.