E2E: Fix form tests for Gutenberg 17.2 changes
In 17.2.0, the "Select X" button in the block toolbar had its text
changed to "Select parent block: X". The test looking for the former
label needs updating for the new one.

In the dialog used for selecting a form pattern, there used to be a bug
in that the a11y label for each option was not on the role=option
element itself. In either 17.2.0 or 17.2.1, that was fixed, so the test
needs updating to no longer try to work around the incorrect structure.