Zendesk: Ensure useChatWidget works without Help Center
In #80075, Help Center stopped proactively loading Messaging's scripts.
Which is a good idea from a performance/resource usage point of view.
However, it broke the cases that use `useChatWidget` (`ChatButton`,
precancelation chat, etc.), because that hook worked under the
assumption that Help Center already loaded the Messaging scripts.

To fix that and make the code more resilient, I've modified
`useChatWidget` to ensure Messaging scripts are loaded. Additionally,
I've removed the reliance on Help Center to show the launcher/widget.
This will make it future-proof for use cases like the one being
implemented in #85381