LOTS: Correct CTA URL for personal tier themes (#86053)
When tiered themes are available `getThemeType` is no longer sufficient
to find the tier of the theme - it treats all themes with a `pub/`
prefix as free themes. With this change, theme tiers is treated as the
source of truth, and mapped to the values expected in onboarding.

Onboarding has been updated so that when starting from the logged-out
theme showcase and choosing a personal-tiered theme, it will show all
plans except the free plan. This is analagous to how premium themes
show all plans except free and personal.